fix: Use auto-discovered organisational preset configuration during on-boarding
When deciding whether to on-board, the global configuration was the only configuration loaded.
This can result in repositories not being on-boarded when the organisational presets being loaded would have resulted in dependencies being identified.
