+++ lib/workers/repository/onboarding/branch/config.spec.ts
@@ -1,7 +1,8 @@
 import { RenovateConfig, getConfig, getName } from '../../../../../test/util';
 import * as presets from '../../../../config/presets/local';
 import { PRESET_DEP_NOT_FOUND } from '../../../../config/presets/util';
-import { getOnboardingConfig } from './config';
+import { RenovateSharedConfig } from '../../../../config/types';
+import { getOnboardingConfig, getOnboardingConfigContents } from './config';
 
 jest.mock('../../../../config/presets/local');
 
@@ -9,21 +10,37 @@ const mockedPresets = presets as jest.Mocked<typeof presets>;
 
 describe(getName(), () => {
   let config: RenovateConfig;
-  let onboardingConfig: string;
+  let onboardingConfig: RenovateSharedConfig;
   beforeEach(() => {
     jest.clearAllMocks();
     config = getConfig();
     config.platform = 'github';
     config.repository = 'some/repo';
   });
+  describe('getOnboardingConfigContents', () => {
+    it('returns the JSON stringified onboarding config', async () => {
+      mockedPresets.getPreset.mockResolvedValueOnce({ enabled: true });
+      const contents = await getOnboardingConfigContents(config);
+      expect(mockedPresets.getPreset).toHaveBeenCalledTimes(1);
+      expect(contents).toEqual(
+        '{\n' +
+          '  "$schema": "https://docs.renovatebot.com/renovate-schema.json",\n' +
+          '  "extends": [\n' +
+          '    "local>some/renovate-config"\n' +
+          '  ]\n' +
+          '}\n'
+      );
+    });
+  });
   describe('getOnboardingConfig', () => {
     it('handles finding an organization preset', async () => {
       mockedPresets.getPreset.mockResolvedValueOnce({ enabled: true });
       onboardingConfig = await getOnboardingConfig(config);
       expect(mockedPresets.getPreset).toHaveBeenCalledTimes(1);
-      expect(JSON.parse(onboardingConfig).extends[0]).toEqual(
-        'local>some/renovate-config'
-      );
+      expect(onboardingConfig).toEqual({
+        $schema: 'https://docs.renovatebot.com/renovate-schema.json',
+        extends: ['local>some/renovate-config'],
+      });
     });
     it('handles finding an organization dot platform preset', async () => {
       mockedPresets.getPreset.mockRejectedValueOnce(
@@ -32,9 +49,10 @@ describe(getName(), () => {
       mockedPresets.getPreset.mockResolvedValueOnce({ enabled: true });
       onboardingConfig = await getOnboardingConfig(config);
       expect(mockedPresets.getPreset).toHaveBeenCalledTimes(2);
-      expect(JSON.parse(onboardingConfig).extends[0]).toEqual(
-        'local>some/.github:renovate-config'
-      );
+      expect(onboardingConfig).toEqual({
+        $schema: 'https://docs.renovatebot.com/renovate-schema.json',
+        extends: ['local>some/.github:renovate-config'],
+      });
     });
     it('handles not finding an organization preset', async () => {
       mockedPresets.getPreset.mockRejectedValue(
@@ -42,7 +60,7 @@ describe(getName(), () => {
       );
       onboardingConfig = await getOnboardingConfig(config);
       expect(mockedPresets.getPreset).toHaveBeenCalledTimes(2);
-      expect(JSON.parse(onboardingConfig)).toEqual(config.onboardingConfig);
+      expect(onboardingConfig).toEqual(config.onboardingConfig);
     });
     it('ignores an unknown error', async () => {
       mockedPresets.getPreset.mockRejectedValue(
@@ -50,7 +68,7 @@ describe(getName(), () => {
       );
       onboardingConfig = await getOnboardingConfig(config);
       expect(mockedPresets.getPreset).toHaveBeenCalledTimes(2);
-      expect(JSON.parse(onboardingConfig)).toEqual(config.onboardingConfig);
+      expect(onboardingConfig).toEqual(config.onboardingConfig);
     });
     it('ignores unsupported platform', async () => {
       mockedPresets.getPreset.mockRejectedValue(
@@ -58,7 +76,7 @@ describe(getName(), () => {
       );
       onboardingConfig = await getOnboardingConfig(config);
       expect(mockedPresets.getPreset).toHaveBeenCalledTimes(2);
-      expect(JSON.parse(onboardingConfig)).toEqual(config.onboardingConfig);
+      expect(onboardingConfig).toEqual(config.onboardingConfig);
     });
   });
 });
