+++ lib/workers/repository/onboarding/branch/index.ts
@@ -1,3 +1,4 @@
+import { mergeChildConfig } from '../../../../config';
 import { getAdminConfig } from '../../../../config/admin';
 import type { RenovateConfig } from '../../../../config/types';
 import {
@@ -8,13 +9,16 @@ import { logger } from '../../../../logger';
 import { platform } from '../../../../platform';
 import { checkoutBranch } from '../../../../util/git';
 import { extractAllDependencies } from '../../extract';
+import { mergeRenovateConfig } from '../../init/merge';
 import { isOnboarded, onboardingPrExists } from './check';
+import { getOnboardingConfig } from './config';
 import { createOnboardingBranch } from './create';
 import { rebaseOnboardingBranch } from './rebase';
 
 export async function checkOnboardingBranch(
-  config: RenovateConfig
+  config_: RenovateConfig
 ): Promise<RenovateConfig> {
+  const config = { ...config_ };
   logger.debug('checkOnboarding()');
   logger.trace({ config });
   const repoIsOnboarded = await isOnboarded(config);
@@ -42,11 +46,18 @@ export async function checkOnboardingBranch(
     }
   } else {
     logger.debug('Onboarding PR does not exist');
-    if (Object.entries(await extractAllDependencies(config)).length === 0) {
+    const onboardingConfig = await getOnboardingConfig(config);
+    let mergedConfig = mergeChildConfig(config, onboardingConfig);
+    mergedConfig = await mergeRenovateConfig(mergedConfig);
+    config.onboardingBranch = mergedConfig.onboardingBranch;
+
+    if (
+      Object.entries(await extractAllDependencies(mergedConfig)).length === 0
+    ) {
       throw new Error(REPOSITORY_NO_PACKAGE_FILES);
     }
     logger.debug('Need to create onboarding PR');
-    const commit = await createOnboardingBranch(config);
+    const commit = await createOnboardingBranch(mergedConfig);
     // istanbul ignore if
     if (commit) {
       logger.info(
